Move propagation core functions
