+++ src/poliastro/tests/test_hyper.py
@@ -4,7 +4,7 @@ import numpy as np
 from numpy.testing import assert_allclose
 from scipy import special
 
-from poliastro.hyper import hyp2f1b as hyp2f1
+from poliastro.core.hyper import hyp2f1b as hyp2f1
 
 
 @pytest.mark.parametrize("x", np.linspace(0, 1, num=11))
