+++ src/poliastro/twobody/propagation.py
@@ -5,15 +5,14 @@ import numpy as np
 import functools
 
 from scipy.integrate import solve_ivp
+
+from poliastro.core.propagation import (
+    mean_motion as mean_motion_fast,
+    kepler as kepler_fast
+)
 from poliastro.integrators import DOP835
 
 from astropy import units as u
-from poliastro.twobody.rv import rv2coe
-from poliastro.twobody.classical import coe2rv
-from poliastro.core.angles import nu_to_M, M_to_nu
-
-from poliastro.core.jit import jit
-from poliastro.stumpff import c2, c3
 
 
 def func_twobody(t0, u_, k, ad, ad_kwargs):
@@ -117,51 +116,7 @@ def mean_motion(orbit, tof, **kwargs):
     r0 = orbit.r.to(u.km).value
     v0 = orbit.v.to(u.km / u.s).value
 
-    return _mean_motion(k, r0, v0, tof)
-
-
-@jit
-def _mean_motion(k, r0, v0, tof):
-    r"""Propagates orbit using mean motion
-
-    .. versionadded:: 0.9.0
-
-    Parameters
-    ----------
-    orbit : ~poliastro.twobody.orbit.Orbit
-        the Orbit object to propagate.
-    tof : float
-        Time of flight (s).
-
-    Notes
-    -----
-    This method takes initial :math:`\vec{r}, \vec{v}`, calculates classical orbit parameters,
-    increases mean anomaly and performs inverse transformation to get final :math:`\vec{r}, \vec{v}`
-    The logic is based on formulae (4), (6) and (7) from http://dx.doi.org/10.1007/s10569-013-9476-9
-
-    """
-
-    # get the initial true anomaly and orbit parameters that are constant over time
-    p, ecc, inc, raan, argp, nu0 = rv2coe(k, r0, v0)
-
-    # get the initial mean anomaly
-    M0 = nu_to_M(nu0, ecc)
-    # strong elliptic or strong hyperbolic orbits
-    if np.abs(ecc - 1.0) > 1e-2:
-        a = p / (1.0 - ecc ** 2)
-        # given the initial mean anomaly, calculate mean anomaly
-        # at the end, mean motion (n) equals sqrt(mu / |a^3|)
-        M = M0 + tof * np.sqrt(k / np.abs(a ** 3))
-        nu = M_to_nu(M, ecc)
-
-    # near-parabolic orbit
-    else:
-        q = p * np.abs(1.0 - ecc) / np.abs(1.0 - ecc ** 2)
-        # mean motion n = sqrt(mu / 2 q^3) for parabolic orbit
-        M = M0 + tof * np.sqrt(k / 2.0 / (q ** 3))
-        nu = M_to_nu(M, ecc)
-
-    return coe2rv(k, p, ecc, inc, raan, argp, nu)
+    return mean_motion_fast(k, r0, v0, tof)
 
 
 def kepler(orbit, tof, *, numiter=350, **kwargs):
@@ -194,7 +149,7 @@ def kepler(orbit, tof, *, numiter=350, **kwargs):
     r0 = orbit.r.to(u.km).value
     v0 = orbit.v.to(u.km / u.s).value
 
-    f, g, fdot, gdot = _kepler(k, r0, v0, tof, numiter)
+    f, g, fdot, gdot = kepler_fast(k, r0, v0, tof, numiter)
 
     assert np.abs(f * gdot - fdot * g - 1) < 1e-5  # Fixed tolerance
 
@@ -211,56 +166,3 @@ def propagate(orbit, time_of_flight, *, method=mean_motion, rtol=1e-10, **kwargs
     """
     r, v = method(orbit, time_of_flight.to(u.s).value, rtol=rtol, **kwargs)
     return orbit.from_vectors(orbit.attractor, r * u.km, v * u.km / u.s, orbit.epoch + time_of_flight)
-
-
-@jit
-def _kepler(k, r0, v0, tof, numiter):
-    # Cache some results
-    dot_r0v0 = np.dot(r0, v0)
-    norm_r0 = np.dot(r0, r0) ** .5
-    sqrt_mu = k**.5
-    alpha = -np.dot(v0, v0) / k + 2 / norm_r0
-
-    # First guess
-    if alpha > 0:
-        # Elliptic orbit
-        xi_new = sqrt_mu * tof * alpha
-    elif alpha < 0:
-        # Hyperbolic orbit
-        xi_new = (np.sign(tof) * (-1 / alpha)**.5 *
-                  np.log((-2 * k * alpha * tof) /
-                  (dot_r0v0 + np.sign(tof) *
-                   np.sqrt(-k / alpha) * (1 - norm_r0 * alpha))))
-    else:
-        # Parabolic orbit
-        # (Conservative initial guess)
-        xi_new = sqrt_mu * tof / norm_r0
-
-    # Newton-Raphson iteration on the Kepler equation
-    count = 0
-    while count < numiter:
-        xi = xi_new
-        psi = xi * xi * alpha
-        c2_psi = c2(psi)
-        c3_psi = c3(psi)
-        norm_r = (xi * xi * c2_psi +
-                  dot_r0v0 / sqrt_mu * xi * (1 - psi * c3_psi) +
-                  norm_r0 * (1 - psi * c2_psi))
-        xi_new = xi + (sqrt_mu * tof - xi * xi * xi * c3_psi -
-                       dot_r0v0 / sqrt_mu * xi * xi * c2_psi -
-                       norm_r0 * xi * (1 - psi * c3_psi)) / norm_r
-        if abs(xi_new - xi) < 1e-7:
-            break
-        else:
-            count += 1
-    else:
-        raise RuntimeError("Maximum number of iterations reached")
-
-    # Compute Lagrange coefficients
-    f = 1 - xi**2 / norm_r0 * c2_psi
-    g = tof - xi**3 / sqrt_mu * c3_psi
-
-    gdot = 1 - xi**2 / norm_r * c2_psi
-    fdot = sqrt_mu / (norm_r * norm_r0) * xi * (psi * c3_psi - 1)
-
-    return f, g, fdot, gdot
